pyproject.toml
[tool.poetry]
name = "metasharks_test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
django-filter = "^21.1"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]



build-backend = "poetry.core.masonry.api"
